pyproject.toml
[project]
name = "test-em"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "bcrypt>=5.0.0",
    "black>=26.1.0",
    "django>=5.2.11",
    "django-filter>=25.2",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "dotenv>=0.9.9",
    "drf-yasg>=1.21.14",
    "psycopg2-binary>=2.9.11",
    "pyjwt>=2.11.0",
    "python-docx>=1.2.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-django>=4.12.0",
]
